backend/requirements.txt
fastapi
uvicorn
python-chess
requests
beautifulsoup4
stockfish
pydantic
python-multipart
sqlalchemy
